requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
llama-cpp-python[cuda]>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
transitions>=0.9.0
httpx>=0.25.0
aiosqlite>=0.19.0
rich>=13.7.0
